+++ CHANGES.md
@@ -10,4 +10,5 @@ in progress
 2023-08-xx 0.0.0
 ================
 
-- Initial thing.
+- Initial thing
+- Journey planning basics and HTML output